Move some Bytes classes from io to bytes package #1199
