+++ src/main/java/org/cactoos/iterable/IterableOfBytes.java
@@ -25,7 +25,7 @@ package org.cactoos.iterable;
 
 import org.cactoos.Bytes;
 import org.cactoos.Text;
-import org.cactoos.io.BytesOf;
+import org.cactoos.bytes.BytesOf;
 import org.cactoos.iterator.IteratorOfBytes;
 
 /**
