+++ src/test/java/org/cactoos/bytes/UncheckedBytesTest.java
@@ -21,53 +21,41 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.io;
+package org.cactoos.bytes;
 
-import java.io.StringReader;
+import java.io.IOException;
+import org.cactoos.Text;
 import org.cactoos.text.TextOf;
-import org.junit.Rule;
+import org.hamcrest.core.IsEqual;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
 import org.llorllale.cactoos.matchers.Assertion;
-import org.llorllale.cactoos.matchers.IsTrue;
-import org.llorllale.cactoos.matchers.TextIs;
 
 /**
- * Test case for {@link ReaderAsBytes}.
+ * Test case for {@link UncheckedBytes}.
  *
- * @since 0.12
+ * @since 0.3
  * @checkstyle JavadocMethodCheck (500 lines)
- * @checkstyle ClassDataAbstractionCouplingCheck (500 lines)
  */
-public final class ReaderAsBytesTest {
-    /**
-     * Temporary files and folders generator.
-     */
-    @Rule
-    public final TemporaryFolder folder = new TemporaryFolder();
+public final class UncheckedBytesTest {
 
-    @Test
-    public void readsString() throws Exception {
-        final String source = "hello, друг!";
-        new Assertion<>(
-            "Must read string through a reader",
-            new TextOf(
-                new ReaderAsBytes(
-                    new StringReader(source)
-                )
-            ),
-            new TextIs(source)
-        ).affirm();
+    @Test(expected = RuntimeException.class)
+    public void rethrowsCheckedToUncheckedException() {
+        new UncheckedBytes(
+            () -> {
+                throw new IOException("intended");
+            }
+        ).asBytes();
     }
 
     @Test
-    public void readsAndClosesReader() throws Exception {
-        final EmptyClosableReader reader = new EmptyClosableReader();
-        new ReaderAsBytes(reader).asBytes();
+    public void worksNormallyWhenNoExceptionIsThrown() throws Exception {
+        final Text source = new TextOf("hello, cactoos!");
         new Assertion<>(
-            "Must close the reader after reading it",
-            reader.isClosed(),
-            new IsTrue()
+            "Must works normally when no exception is thrown",
+            new UncheckedBytes(
+                new BytesOf(source)
+            ).asBytes(),
+            new IsEqual<>(new BytesOf(source).asBytes())
         ).affirm();
     }
 }
