+++ src/test/java/org/cactoos/io/LoggingOutputTest.java
@@ -28,6 +28,7 @@ import java.io.OutputStream;
 import java.nio.file.Path;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import org.cactoos.bytes.BytesOf;
 import org.cactoos.scalar.LengthOf;
 import org.hamcrest.Matchers;
 import org.junit.Rule;
